pyproject.toml
[project]
name = "blog-to-podcast"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "elevenlabs>=2.8.1",
    "google-generativeai==0.8.2",
    "gtts>=2.5.4",
    "pydub>=0.25.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "ruff>=0.12.7",
]
